Move solicitor executor transformer into directory
